Emit warning on unexpected tag instead of crashing
Sometimes a package is incorrectly pushed to multiple repositories of different
versions (e.g. a CentOS 5 and CentOS 6 repository). When attepting a pullpackage
to remove it from the wrong repository, it crashes during dependency testing
preprocessing because it expects that all of the tags belong to the same CentOS
version as the repository being pulled from.

This fix causes pullpackage to emit a warning when it detects an issue but it
will no longer fail on exception.